requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.2.35
cryptography==39.0.2
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-import-export==3.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
MarkupPy==1.14
msgpack==1.0.5
odfpy==1.4.1
openpyxl==3.1.2
prompt-toolkit==3.0.38
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.3
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
